add adaptive, change logout button

+++ src/layouts/MainLayout.jsx
@@ -8,6 +8,7 @@ import { Link, useNavigate } from "react-router-dom";
 import {
   MenuFoldOutlined,
   MenuUnfoldOutlined,
+  PoweroffOutlined,
   UserOutlined
 } from '@ant-design/icons';
 import { Layout, Menu, Button, theme } from 'antd';
@@ -20,56 +21,87 @@ import "../pages/AccountPage/AccountPage";
 
 const { Header, Sider, Content } = Layout;
 
-const NAVIGATION_ITEMS = [
 
-  {
+
+
+function MainLayout({ children }) {
+
+  const NAVIGATION_ITEMS = [{
     label: <Link to="/account">My account</Link>,
     key: "/account",
     icon: <UserOutlined />,
+    name: "My account"
   },
   {
     label: <Link to="/calendar">Calendar</Link>,
     key: "/calendar",
-    icon: <BsCalendar2Check />
+    icon: <BsCalendar2Check />,
+    name: "Calendar"
   },
+
+  {
+    label: <Link>Logout</Link>,
+    key: "/logout",
+    icon: <PoweroffOutlined />,
+    danger: true,
+    name: "Logout"
+  },
+  
 ];
 
-
-function MainLayout({ children }) {
   const [collapsed, setCollapsed] = useState(false);
+
+
+  const path = window.location.pathname;
+  console.log(path);
+  const startMenuItemName = NAVIGATION_ITEMS.find(x => x.key === path).name;
+
+  const [menuItemName, setMenuItemName] = useState(startMenuItemName);
   const {
     token: { colorBgContainer },
   } = theme.useToken();
 
+  const setHeaderDescription = (e) => {
+    console.log(e.item.props.name);
+    setMenuItemName(e.item.props.name);
+  }
 
-  const handleLogoutClick = e => {
-    const result = window.confirm('Do you really want to log out?');
-    if (result) {
-      localStorage.clear();
-      navigate("/login");
+  const getUserName = () => {
+    const stringFromLocalstorate = localStorage.getItem("auth");
 
+    if(stringFromLocalstorate) {
+      const obj = JSON.parse(stringFromLocalstorate);
+      
+      return obj?.user?.name;
     }
-  };
+
+    return '';
+  }
 
 
   const navigate = useNavigate();
 
-  const handleLogoClick = e => {
+  const handleLogoClick = (e) => {
     {
       navigate("/calendar")
+    }
+  };
+
+  const handleMenuClick = (info) => { 
+    console.log(info);
+    if(info.key === "/logout") {
+        let result = window.confirm('Do you really want to log out?');
+        if (result) {
+          localStorage.clear();  
+          navigate("/login");
+        }
     }
   };
 
   return (
     <Layout className="main-block">
       <Sider trigger={null} collapsible collapsed={collapsed}  
-      // style={{
-      //     display: 'block',
-      //     '@media (max-width: 768px)': {
-      //       display: 'none',
-      //     },
-      //   }}
-        >
+      >
         <div className="sider">
           <svg onClick={handleLogoClick} className="logo-svg" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 48 48" width="48px" height="48px"><linearGradient id="IMoH7gpu5un5Dx2vID39Ra" x1="9.858" x2="38.142" y1="9.858" y2="38.142" gradientUnits="userSpaceOnUse"><stop offset="0" stopColor="#9dffce" /><stop offset="1" stopColor="#50d18d" /></linearGradient><path fill="url(#IMoH7gpu5un5Dx2vID39Ra)" d="M44,24c0,11.045-8.955,20-20,20S4,35.045,4,24S12.955,4,24,4S44,12.955,44,24z" /><linearGradient id="IMoH7gpu5un5Dx2vID39Rb" x1="13" x2="36" y1="24.793" y2="24.793" gradientUnits="userSpaceOnUse"><stop offset=".824" stopColor="#135d36" /><stop offset=".931" stopColor="#125933" /><stop offset="1" stopColor="#11522f" /></linearGradient><path fill="url(#IMoH7gpu5un5Dx2vID39Rb)" d="M21.293,32.707l-8-8c-0.391-0.391-0.391-1.024,0-1.414l1.414-1.414	c0.391-0.391,1.024-0.391,1.414,0L22,27.758l10.879-10.879c0.391-0.391,1.024-0.391,1.414,0l1.414,1.414	c0.391,0.391,0.391,1.024,0,1.414l-13,13C22.317,33.098,21.683,33.098,21.293,32.707z" /></svg>
           <h1
@@ -80,26 +112,32 @@ function MainLayout({ children }) {
           <div className="demo-logo-vertical" />
         </div>
         <Menu
+          onClick={(info) => handleMenuClick(info)}
           items={NAVIGATION_ITEMS}
           theme="dark"
           mode="inline"
           defaultSelectedKeys={[window.location.pathname]}
+          onSelect={e => setHeaderDescription(e)}
         />
-        <Button
-          onClick={handleLogoutClick}
+        {/* <Button
+          onClick={handleLogoutAndToggleSidebar}
           className="sider-logout"
-        ><span className="general-font">Logout</span>
-          <RiLogoutBoxRLine />
-        </Button>
+          style={{ width: isLogoutButtonCollapsed ? '64' : '180',
+          position: isLogoutButtonCollapsed ? 'absolute' : 'absolut', // Установите позицию кнопки в зависимости от collapsed
+          right: isLogoutButtonCollapsed ? '8' : '33'
+         }}
+        ><span className="general-font"><PoweroffOutlined /> Logout</span>
+          
+        </Button> */}
       </Sider>
 
       <Layout>
         <Header className="header" style={{ padding: 0, background: colorBgContainer }}>
           <div className="header-wrapper">
-            <h1 className="header-page">My account</h1>
+            <h1 className="header-page">{menuItemName}</h1>
             <div className="header-block">
               <MdNightlight className="header-block__theme" />
-              <h2 className="header-block__name">userName</h2>
+              <h2 className="header-block__name">{getUserName()}</h2>
               <FaUserCircle
                 className="header-block__icon"
                 onClick={() => navigate("/account")}
@@ -114,6 +152,8 @@ function MainLayout({ children }) {
               fontSize: '16px',
               width: 64,
               height: 64,
+              position: 'absolute',
+              top: '-11px'
             }}
           />
